+++ packages/coder/src/collecter.ts
@@ -147,7 +147,7 @@ export class Collecter {
   private collectStyle(node: NodeSchema) {
     if (node.id && node.props?.style) {
       const hasStyle = !!Object.keys(node.props.style).length;
-      if (hasStyle) {
+      if (hasStyle && !isJSCode(node.props.style)) {
         this.style[`.${node.name}_${node.id}`] = node.props.style;
       }
     }