+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.10](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.11.9...@vtj/ui@0.11.10) (2025-03-27)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.11.9](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.11.8...@vtj/ui@0.11.9) (2025-03-27)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.11.8](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.11.7...@vtj/ui@0.11.8) (2025-03-27)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.11.7](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.11.6...@vtj/ui@0.11.7) (2025-03-25)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
+## [0.11.6](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.11.5...@vtj/ui@0.11.6) (2025-03-21)
+
+**Note:** Version bump only for package @vtj/ui
+
+
+
+
+
 ## [0.11.5](https://gitee.com/newgateway/vtj/compare/@vtj/ui@0.11.4...@vtj/ui@0.11.5) (2025-03-20)
 
 **Note:** Version bump only for package @vtj/ui