+++ platforms/pro/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.10](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.11.9...@vtj/pro@0.11.10) (2025-03-27)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.11.9](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.11.8...@vtj/pro@0.11.9) (2025-03-27)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.11.8](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.11.7...@vtj/pro@0.11.8) (2025-03-27)
+
+
+### Features
+
+* ✨ report 模块,修复access和style出码bug ([db401ee](https://gitee.com/newgateway/vtj/commits/db401ee22a2c2fb85f4867755047966164b8e9f8))
+
+
+
+
+
+## [0.11.7](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.11.6...@vtj/pro@0.11.7) (2025-03-25)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.11.6](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.11.5...@vtj/pro@0.11.6) (2025-03-21)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
 ## [0.11.5](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.11.4...@vtj/pro@0.11.5) (2025-03-20)
 
 **Note:** Version bump only for package @vtj/pro